twinBASIC bug - VB.Picturebox and VB.VBScrollBar
If you have a form where you create a picturebox or a vbscrollbar dynamically the built exe will not display them. The work around is you need to add either to the form. For speed I just placed invisible ones on the forms using them, so as not to have to change the code behind as well.