Update Form1.twin
Strangely, tB did not mark this line as an error when I removed that procedure from ucCustomButton. It only showed the lines in a different form (frmShowMessage) I even ran the code in the IDE. It was found after attempting to build an EXE and it showed after I closed and reopened the project.