beta 2 - 10-19-25
Adding DPI code to Activity Log and Change Log
updated columns width calculation, type badge drawing, type text location in the badge, row height, and inner row wrapping.

Add code given by ArtAraya, again thank you, to handle the custom buttons, as well as what needed to be changed for the images (used for placing icons) to be correct.

Modified selecting a twinBASIC location code as it lacked adding "\" causing the check for the existence of tB in the folder to break. 
Added the verification that the download button can be enabled once both the folders are valid.